More scalable organization of files
- Moved the lambda functions and tests to be better organized.
- Added appropriate response objects for functions when triggered.
- Simplified parsing of the username from an ARN
