+++ functions/tests/test_lacework_event_router.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Tests for Lacework Event Router.
+"""
+
+pytest_plugins = [
+    "tests.iam.test_iam",
+]
